Implement a simple debug log formatting utility
This change includes a new enviroment.h file that is planned to be
used to include all the compiler and platform specific macros.
Specificly, this change includes the definition of the 'weak' macro
that is used to define a function that can be overriden by the
platform specific implementation.

The debug log utility is implemented in libe15-dbg.c and libe15-dbg.h
and is used to format the debug log messages. data output is done
using a weak function 'dev_putc' that should be implemented by the
user.