Add： code framework for modbus protocol.
Modbus - RTU protocol is a serial communication protocol, which is
used to communicate with the device through the serial port.

This implementation require user to implement it's own serial port
driver, and provide the serial port driver to the modbus driver by
calling `modbus_slave_recv_handler` and implement `set_send_buffer`
function.

This commit is a framework for this implementation, currently no
function is implemented.